+++ src/components/custom_footer/styles.js
@@ -1,0 +1,21 @@
+import {StyleSheet} from 'react-native';
+import {resize} from '../../utils/styles';
+import colors from '../../utils/colors';
+
+export default StyleSheet.create({
+  footer: {
+    backgroundColor: colors.white,
+  },
+  icon: {
+    color: colors.colorText,
+  },
+  iconActive: {
+    color: colors.titleColorText,
+  },
+  text: {
+    color: colors.colorText,
+  },
+  textActive: {
+    color: colors.titleColorText,
+  },
+});