+++ src/components/header/index.js
@@ -3,9 +3,23 @@ import {SafeAreaView} from 'react-native';
 
 import CustomHeader from '../custom_headers';
 
-const Header = ({type, title}) => (
+const Header = ({
+  type,
+  title,
+  searchValue,
+  onLeft,
+  onSearchChange,
+  clearSearch,
+}) => (
   <SafeAreaView>
-    <CustomHeader title={title} type={type} />
+    <CustomHeader
+      searchValue={searchValue}
+      clearSearch={clearSearch}
+      onSearchChange={onSearchChange}
+      onLeft={() => onLeft()}
+      title={title}
+      type={type}
+    />
   </SafeAreaView>
 );