+++ src/routes/app/index.js
@@ -5,13 +5,13 @@ import {createStackNavigator} from 'react-navigation-stack';
 // screens
 import Home from '../../screens/app/Home';
 import DetailData from '../../screens/app/DetailData';
+import Search from '../../screens/app/Search';
 
 const appStackNavigator = createStackNavigator(
   {
-    Home: {
-      screen: Home,
-    },
+    Home,
     DetailData,
+    Search,
   },
   {
     initialRouteName: 'Home',